Assert that weights and gradWeights line up in getParameters
Adds a check that the parameters have the same offset as their gradients
after getParameters is called. If they do not line up, then methods such
as torch/optim will not work. This could happen if the sharing of
weights and gradWeights do not match or, due to a bug in the
implementation of getParameters, if the storages of weights and
gradWeights do not closely correspond.

Fix getParameters tests to always share gradWeights when sharing
weights.