[DistKLDivCriterion] Add assertions on target size
In contrast to many other criterions, `DistKLDivCriterion` requires the size of `target` to be the same as that of `input` (for obvious reasons). In case you do this wrong (which can easily happen), there is currently no informative error message but only a failure in the CUDA-code. This diff adds assertions on the `input` and `target` size.