Don't serialize shared parameters in clone()
There's no need to serialize parameters that will be shared immediately
afterwards. In particular, the current implementation of clone and share is
inefficient for single modules in a larger network with flattened parameters
since the storage holding all parameters will be serialized.

Here, the write method of nn.Module is temporarily overwritten with a function
that writes empty tensors for all shared parameters.