Support index parameters with Cuda*Tensor.
These modules check the type explicitly for CudaTensor
rather than the regular expression.  There are other
modules that have this identical check, but do operations
on the resulting type which may or may not be support
(i.e. require more checking), so I didn't change them.
These should be safe because the index type is independent
of the type being checking.